pyproject.toml
[build-system]
requires = [
    "setuptools>=80.0.0",
    "setuptools_scm[toml]>=8",
]
build-backend = "setuptools.build_meta"

[project]
name = "card_bill_collector"
description = "Getting information from a bank account credit card bill and expense analysis"
readme = "README.md"
authors = [{ name = "Felipe Getúlio", email = "felipegetulio15@gmail.com" }]
license = { file = "LICENSE" }
requires-python = ">=3.10"
dynamic = ["version"]
dependencies = [
    "python-dotenv",
    "uiautomator2==3.3.3",
    "py-moneyed",
    "ordered-set",
    "pandas",
    "dateparser"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "no-local-version"
